Renamed "DiscreteDefinition" to "DiscreteRepresentation" in schema, and the correcponding attribute in Problem as well.
Updated the code and tests to reflect this change. Also added the attribute `non_dominated` to DiscreteRepresenation. This
is useful information in many places. DiscreteRepresenation reflects the purpose of the field better. It can also be used for
defining problems for methods that require both an analytical and discrete representation of the problem.