Problem schema and documentation changes.
The EvlauatedSolutions and EvaluatedInfo models have been removed from now. They were not used anywhere and are not really part of defining a problem. A similar model will be useful later on the emo side, probably, but right now, it is just confusing. DiscreteRepresentation does the job better.

Rewrote and restuctred the documentation on the Problem format. Will continue writing on problem parsing and evaluating, solvers, and scalarization later.