Differentiable variant of NIMBUS scalarization + schema.
Implemented the differentiable variant of the NIMBUS scalarization
function. Tested its initialization, need to test the correct
calculation of Pareto optimal solutions next. Some getters implemented
in the Problem schema as well, nice utilities. Changed some headrers in
the reference markdown files in the docs, did not get the need for the
change, was confusing. This is, however, debatable.