Added script to create two types of requirements.
The bash script 'make_requirements.sh' can be given either the argument
'rtd', in which case it creates a readthedocs.org suitable requirements.txt
file; or the argument 'pip', in which case it creates a requirements.txt
file suitable for installing the project using pip/pipx. Check the
file for more details.