Intermediate progress on the nevergrad solver.
Nevergrad does not work well with vectorized function calls. I think the
underlying assumtion is to call a function once, adjust the model, call
function again, and so forth. This is ideal for heavy black-box
functionns. However, to evaluate analytic and other functions, we need a
parser that provides fast one-shot evaluation. For this, a sympy parser
is probably the easiest approach.