Added one more NIMBUS test.
Tried to replicate the NIMBUS papre results, but could not. The first
result calcualted with the NIMBUS scalarization function is the same as
in the paper up to a precision of 1e-2, but the results computed with
the other scalarization functions vary wildly. Witout access to the
original implementation, this comparison is not very meaningful.