Random access.
- Infix parser now supports random access of (tensor) variables using
bracket, e.g., X[1,2] -> ["At", X, 1 ,2]. This is similar to how random
access is defined in MathJSON.

- Implemented random access to the Pyomo math parser. Tested.
Seems to work just fine.

- Implemented __getitem__ for TensorConstant and tested it. It returns a
new Constantn that has been initialized with the information found at
the given position in TensorConstant, e.g., the current constnat values.
Name and symbol are also derived from TensorConstnat based on the
indices given to __getitem__.

- Implemented random access for TensorVariable. It returns a copy of
Variable with the TensorVariable's data at position TensorVariable[x,
y].

- Implemented the 'to_contants' and 'to_variables' methods for
TensorConstant and TensorVariable, respectively. These flatten the
tensors and return a list of Constants or Variables, depending whether
the tensor was a TensorConstant or TensorVariable.